test: share Telegram media resolution fixtures
